IzoterikBackend/Backend/pyproject.toml
[project]
name = "Backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Daniil Ermolaev", email = "twofaced-janus@yandex.ru"},
]
dependencies = ["django>=5.2.7", "djangorestframework>=3.16.1", "markdown>=3.9", "django-filter>=25.2", "psycopg2-binary>=2.9.9", "djangorestframework-simplejwt>=5.3.0", "rest-framework-simplejwt>=0.0.2", "django-cors-headers>=4.9.0"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
